@@ -1,44 +1,52 @@
 {
   "name": "metamodels/notelist",
-  "description": "Notelist management for MetaModels",
+  "description": "Note list management for MetaModels",
   "license": "LGPL-3.0-or-later",
-  "keywords":["contao", "metamodels", "notelist"],
-  "type":"contao-module",
-  "homepage":"http://now.metamodel.me/",
-  "authors":[
+  "type": "contao-bundle",
+  "keywords": [
+    "contao",
+    "metamodels",
+    "notelist",
+    "note list"
+  ],
+  "authors": [
     {
-      "name":"Christian Schiffler",
-      "email":"c.schiffler@cyberspectrum.de",
-      "homepage":"http://www.cyberspectrum.de",
-      "role":"Developer"
+      "name": "Christian Schiffler",
+      "email": "c.schiffler@cyberspectrum.de",
+      "homepage": "https://www.cyberspectrum.de",
+      "role": "Developer"
     },
     {
-      "name":"Ingolf Steinhardt",
-      "email":"info@e-spin.de",
-      "homepage":"http://www.e-spin.de",
-      "role":"Developer"
+      "name": "Ingolf Steinhardt",
+      "email": "info@e-spin.de",
+      "homepage": "https://www.e-spin.de",
+      "role": "Developer"
     }
   ],
-  "support":{
-    "email":"mail@metamodel.me",
-    "issues":"https://github.com/MetaModels/metamodels_notelist",
-    "wiki":"http://de.contaowiki.org/MetaModels",
-    "irc":"irc://irc.freenode.org/contao.mm",
-    "source":"https://github.com/MetaModels/metamodels_notelist"
+  "homepage": "https://now.metamodel.me/",
+  "support": {
+    "email": "mail@metamodel.me",
+    "issues": "https://github.com/MetaModels/metamodels_notelist",
+    "wiki": "https://metamodels.readthedocs.io/de/latest/manual/extended/notelist.html",
+    "source": "https://github.com/MetaModels/metamodels_notelist"
   },
-  "require":{
-    "php": "^7.0",
-    "contao-community-alliance/dc-general": "^2.1",
-    "contao-community-alliance/url-builder": "^1.3",
-    "contao/core-bundle": "^4.4",
-    "menatwork/contao-multicolumnwizard-bundle": "^3.4",
-    "metamodels/core":"^2.1",
-    "symfony/event-dispatcher": "^2.8 | ^3.0 | ^4.0"
+  "require": {
+    "php": "^8.1",
+    "contao-community-alliance/dc-general": "^2.3",
+    "contao-community-alliance/url-builder": "^1.3.3",
+    "contao/core-bundle": "^4.13.0 <5.0",
+    "menatwork/contao-multicolumnwizard-bundle": "^3.6.7",
+    "metamodels/core": "^2.3",
+    "symfony/dependency-injection": "^5.4",
+    "symfony/http-foundation": "^5.4",
+    "symfony/http-kernel": "^5.4",
+    "symfony/translation": "^5.4",
+    "symfony/translation-contracts": "^2.5.2"
   },
   "require-dev": {
-    "cyberspectrum/contao-toolbox": "^0.7",
-    "phpcq/all-tasks": "^1.2",
-    "contao/manager-plugin": "^2.1"
+    "contao/manager-plugin": "^2.8",
+    "dg/bypass-finals": "^1.9",
+    "phpcq/runner-bootstrap": "^1.0@dev"
   },
   "autoload": {
     "psr-4": {
@@ -53,18 +61,18 @@
       "MetaModels\\NoteListBundle\\Test\\": "tests"
     }
   },
-  "extra":{
-    "contao-manager-plugin": "MetaModels\\NoteListBundle\\ContaoManager\\Plugin",
-    "contao": {
-      "transifex": {
-        "project": "metamodels",
-        "prefix": "notelist-",
-        "languages_cto": "contao/languages",
-        "languages_tx": ".tx"
-      }
+  "config": {
+    "allow-plugins": {
+      "contao-components/installer": true,
+      "contao/manager-plugin": true,
+      "php-http/discovery": true
     },
+    "sort-packages": true
+  },
+  "extra": {
     "branch-alias": {
-      "dev-hotfix/2.1.2": "2.1.2-dev"
-    }
+      "dev-feature/2.3.0": "2.3.x-dev"
+    },
+    "contao-manager-plugin": "MetaModels\\NoteListBundle\\ContaoManager\\Plugin"
   }
 }
