@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,21 +12,25 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\Bridge;
 
 use Contao\DataContainer;
+use Contao\System;
 use Doctrine\DBAL\Connection;
 use MetaModels\BackendIntegration\TemplateList;
 use MetaModels\IFactory;
+use MetaModels\IMetaModel;
 use MetaModels\NoteListBundle\NoteListFactory;
 use MenAtWork\MultiColumnWizardBundle\Contao\Widgets\MultiColumnWizard;
+use Symfony\Component\DependencyInjection\ServiceLocator;
 
 /**
  * This class bridges Contao callbacks to proper listeners.
@@ -40,17 +44,26 @@ class DcaCallbackBridge
      *
      * @return array
      */
-    public static function getNoteListOptions(DataContainer $dataContainer)
+    public static function getNoteListOptions(DataContainer $dataContainer): array
     {
         $container       = self::getLocator();
         $factory         = $container->get(IFactory::class);
         $noteListFactory = $container->get(NoteListFactory::class);
 
-        $metaModelId = $dataContainer->activeRecord->metamodel;
-        $metaModel   = $factory->getMetaModel($factory->translateIdToMetaModelName($metaModelId));
-        if (!$metaModel) {
+        /** @psalm-suppress UndefinedMagicPropertyFetch */
+        if (
+            null === $dataContainer->activeRecord
+            || !(\is_string($metaModelId = $dataContainer->activeRecord->metamodel) || \is_int($metaModelId))
+            || empty($metaModelId)
+        ) {
             return [];
         }
+
+        $metaModel = $factory->getMetaModel($factory->translateIdToMetaModelName((string) $metaModelId));
+        if (null === $metaModel) {
+            return [];
+        }
+
         return $noteListFactory->getConfiguredListsFor($metaModel);
     }
 
@@ -59,7 +72,7 @@ class DcaCallbackBridge
      *
      * @return array
      */
-    public static function getMetaModelOptions()
+    public static function getMetaModelOptions(): array
     {
         $container = self::getLocator();
         $factory   = $container->get(IFactory::class);
@@ -68,7 +81,9 @@ class DcaCallbackBridge
 
         $result = [];
         foreach ($metaModels as $metaModel) {
-            $instance = $factory->getMetaModel($metaModel);
+            if (!(($instance = $factory->getMetaModel($metaModel)) instanceof IMetaModel)) {
+                continue;
+            }
 
             $result[$instance->get('id')] = $instance->getName();
         }
@@ -83,17 +98,29 @@ class DcaCallbackBridge
      *
      * @return string[]
      */
-    public static function getNoteListOptionsMcw(MultiColumnWizard $wizard)
+    public static function getNoteListOptionsMcw(MultiColumnWizard $wizard): array
     {
         $container       = self::getLocator();
         $factory         = $container->get(IFactory::class);
         $noteListFactory = $container->get(NoteListFactory::class);
 
-        $metaModelId = $wizard->activeRecord->metamodel;
-        $metaModel   = $factory->getMetaModel($factory->translateIdToMetaModelName($metaModelId));
-        if (!$metaModel) {
+        /**
+         * @psalm-suppress DocblockTypeContradiction
+         * @psalm-suppress UndefinedMagicPropertyFetch
+         */
+        if (
+            null === $wizard->activeRecord
+            || !(\is_string($metaModelId = $wizard->activeRecord->metamodel) || \is_int($metaModelId))
+            || empty($metaModelId)
+        ) {
             return [];
         }
+
+        $metaModel = $factory->getMetaModel($factory->translateIdToMetaModelName((string) $metaModelId));
+        if (null === $metaModel) {
+            return [];
+        }
+
         return $noteListFactory->getConfiguredListsFor($metaModel);
     }
 
@@ -104,8 +131,20 @@ class DcaCallbackBridge
      *
      * @return string[] array of all attributes as id => human name
      */
-    public function getRenderSettingsMcw(MultiColumnWizard $wizard)
+    public function getRenderSettingsMcw(MultiColumnWizard $wizard): array
     {
+        /**
+         * @psalm-suppress DocblockTypeContradiction
+         * @psalm-suppress UndefinedMagicPropertyFetch
+         */
+        if (
+            null === $wizard->activeRecord
+            || !(\is_string($metaModelId = $wizard->activeRecord->metamodel) || \is_int($metaModelId))
+            || empty($metaModelId)
+        ) {
+            return [];
+        }
+
         /** @var Connection $database */
         $database = $this->getLocator()->get(Connection::class);
 
@@ -114,17 +153,18 @@ class DcaCallbackBridge
             ->select('*')
             ->from('tl_metamodel_rendersettings')
             ->where('pid=:pid')
-            ->setParameter('pid', $wizard->activeRecord->metamodel)
-            ->execute()
-            ->fetchAll(\PDO::FETCH_ASSOC);
+            ->setParameter('pid', (int) $metaModelId)
+            ->executeQuery()
+            ->fetchAllAssociative();
 
-        $result = array();
+        $result = [];
         foreach ($renderSettings as $renderSetting) {
             $result[$renderSetting['id']] = $renderSetting['name'];
         }
 
         // Sort the render settings.
-        asort($result);
+        \asort($result);
+
         return $result;
     }
 
@@ -143,10 +183,13 @@ class DcaCallbackBridge
     /**
      * Obtain the service locator.
      *
-     * @return \Symfony\Component\DependencyInjection\ServiceLocator
+     * @return ServiceLocator
      */
-    private static function getLocator()
+    private static function getLocator(): ServiceLocator
     {
-        return \Contao\System::getContainer()->get('metamodels-notelist.bridge-locator');
+        $serviceLocator = System::getContainer()->get('metamodels-notelist.bridge-locator');
+        assert($serviceLocator instanceof ServiceLocator);
+
+        return $serviceLocator;
     }
 }
