@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2023 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,19 +12,20 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2023 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\Event;
 
 use MetaModels\IItem;
 use MetaModels\IMetaModel;
 use MetaModels\NoteListBundle\Storage\NoteListStorage;
-use Symfony\Component\EventDispatcher\Event;
+use Symfony\Contracts\EventDispatcher\Event;
 
 /**
  * This event gets triggered when a note list gets manipulated.
@@ -34,45 +35,45 @@ class ManipulateNoteListEvent extends Event
     /**
      * An item has been added.
      */
-    const OPERATION_ADD = 'add';
+    public const OPERATION_ADD = 'add';
 
     /**
      * An item has been removed.
      */
-    const OPERATION_REMOVE = 'remove';
+    public const OPERATION_REMOVE = 'remove';
 
     /**
      * The list has been cleared.
      */
-    const OPERATION_CLEAR = 'clear';
+    public const OPERATION_CLEAR = 'clear';
 
     /**
      * The MetaModel.
      *
      * @var IMetaModel
      */
-    private $metaModel;
+    private IMetaModel $metaModel;
 
     /**
      * The note list being manipulated.
      *
      * @var NoteListStorage
      */
-    private $noteList;
+    private NoteListStorage $noteList;
 
     /**
      * The operation.
      *
      * @var string
      */
-    private $operation;
+    private string $operation;
 
     /**
      * The item being added/removed (if any).
      *
      * @var IItem|null
      */
-    private $item;
+    private ?IItem $item;
 
     /**
      * Create a new instance.
@@ -80,7 +81,7 @@ class ManipulateNoteListEvent extends Event
      * @param IMetaModel      $metaModel The MetaModel.
      * @param NoteListStorage $noteList  The note list.
      * @param string          $operation The operation.
-     * @param IItem           $item      The item being added/removed (if any).
+     * @param IItem|null      $item      The item being added/removed (if any).
      */
     public function __construct(IMetaModel $metaModel, NoteListStorage $noteList, string $operation, IItem $item = null)
     {
