@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,18 +12,20 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\EventListener\DcGeneral;
 
 use ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\Event\GetPropertyOptionsEvent;
-use ContaoCommunityAlliance\Translator\TranslatorInterface;
+use ContaoCommunityAlliance\DcGeneral\DataDefinition\ContainerInterface;
 use MetaModels\NoteListBundle\Storage\StorageAdapterFactory;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * This class provides the list of registered storage adapters for the backend.
@@ -35,14 +37,14 @@ class AdapterListListener
      *
      * @var StorageAdapterFactory
      */
-    private $factory;
+    private StorageAdapterFactory $factory;
 
     /**
      * The translator.
      *
      * @var TranslatorInterface
      */
-    private $translator;
+    private TranslatorInterface $translator;
 
     /**
      * Create a new instance.
@@ -70,15 +72,19 @@ class AdapterListListener
             return;
         }
 
-        if (('storageAdapter' !== $event->getPropertyName())
-        || ('tl_metamodel_notelist' !== $event->getEnvironment()->getDataDefinition()->getName())) {
+        if (
+            ('storageAdapter' !== $event->getPropertyName())
+            || !(($dataDefinition = $event->getEnvironment()->getDataDefinition()) instanceof ContainerInterface)
+            || ('tl_metamodel_notelist' !== $dataDefinition->getName())
+        ) {
             return;
         }
 
         $adapters = $this->factory->getIdentifiers();
         $result   = [];
         foreach ($adapters as $adapter) {
-            $result[$adapter] = $this->translator->translate('adapter.' . $adapter, 'metamodels_notelist');
+            $result[$adapter] =
+                $this->translator->trans('notelist.adapter.' . $adapter, [], 'notelist_default');
         }
 
         $event->setOptions($result);
