@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017-2018 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,22 +13,25 @@
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2017-2018 The MetaModels team.
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\EventListener\DcGeneral;
 
 use ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\Event\BuildWidgetEvent;
 use ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\Event\DecodePropertyValueForWidgetEvent;
 use ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\Event\EncodePropertyValueFromWidgetEvent;
+use ContaoCommunityAlliance\DcGeneral\DataDefinition\ContainerInterface;
 use ContaoCommunityAlliance\DcGeneral\Data\ModelInterface;
-use ContaoCommunityAlliance\Translator\TranslatorInterface;
+use Contao\StringUtil;
+use MetaModels\Dca\Helper;
 use MetaModels\IFactory;
 use MetaModels\IMetaModel;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * This class handles the building of the MCW for note list names.
@@ -40,14 +43,14 @@ class BuildNoteListNameWidgetListener
      *
      * @var IFactory
      */
-    private $factory;
+    private IFactory $factory;
 
     /**
      * The translator.
      *
      * @var TranslatorInterface
      */
-    private $translator;
+    private TranslatorInterface $translator;
 
     /**
      * Create a new instance.
@@ -61,7 +64,6 @@ class BuildNoteListNameWidgetListener
         $this->translator = $translator;
     }
 
-
     /**
      * Decode the given value from a serialized language array into the real language array.
      *
@@ -69,39 +71,19 @@ class BuildNoteListNameWidgetListener
      *
      * @return void
      */
-    public function decodeNameValue(DecodePropertyValueForWidgetEvent $event)
+    public function decodeNameValue(DecodePropertyValueForWidgetEvent $event): void
     {
-        if (('tl_metamodel_notelist' !== $event->getModel()->getProviderName())
-            || ($event->getProperty() !== 'name')) {
+        if (
+            ($event->getProperty() !== 'name')
+            || ('tl_metamodel_notelist' !== $event->getModel()->getProviderName())
+        ) {
             return;
         }
 
         $metaModel = $this->getMetaModelByModelPid($event->getModel());
-        $value     = deserialize($event->getValue());
-        if (!$metaModel->isTranslated()) {
-            // If we have an array, return the first value and exit, if not an array, return the value itself.
-            $event->setValue(is_array($value) ? $value[key($value)] : $value);
-            return;
-        }
+        $values    = Helper::decodeLangArray($event->getValue(), $metaModel);
 
-        // Sort like in MetaModel definition.
-        $output = [];
-        if (!empty($languages = $metaModel->getAvailableLanguages())) {
-            foreach ($languages as $strLangCode) {
-                if (is_array($value)) {
-                    $subValue = $value[$strLangCode];
-                } else {
-                    $subValue = $value;
-                }
-
-                if (is_array($subValue)) {
-                    $output[] = array_merge($subValue, ['langcode' => $strLangCode]);
-                } else {
-                    $output[] = ['langcode' => $strLangCode, 'value' => $subValue];
-                }
-            }
-        }
-        $event->setValue(serialize($output));
+        $event->setValue($values);
     }
 
     /**
@@ -111,31 +93,19 @@ class BuildNoteListNameWidgetListener
      *
      * @return void
      */
-    public function encodeNameValue(EncodePropertyValueFromWidgetEvent $event)
+    public function encodeNameValue(EncodePropertyValueFromWidgetEvent $event): void
     {
-        if (('tl_metamodel_notelist' !== $event->getModel()->getProviderName())
-            || ($event->getProperty() !== 'name')) {
+        if (
+            ($event->getProperty() !== 'name')
+            || ('tl_metamodel_notelist' !== $event->getModel()->getProviderName())
+        ) {
             return;
         }
 
         $metaModel = $this->getMetaModelByModelPid($event->getModel());
-        // Not translated, nothing to do.
-        if (!$metaModel->isTranslated()) {
-            return;
-        }
-        $output = [];
+        $values    = Helper::encodeLangArray($event->getValue(), $metaModel);
 
-        foreach (deserialize($event->getValue()) as $subValue) {
-            $langcode = $subValue['langcode'];
-            unset($subValue['langcode']);
-            if (count($subValue) > 1) {
-                $output[$langcode] = $subValue;
-            } else {
-                $keys              = array_keys($subValue);
-                $output[$langcode] = $subValue[$keys[0]];
-            }
-        }
-        $event->setValue(serialize($output));
+        $event->setValue($values);
     }
 
     /**
@@ -145,80 +115,33 @@ class BuildNoteListNameWidgetListener
      *
      * @return void
      */
-    public function buildWidget(BuildWidgetEvent $event)
+    public function buildWidget(BuildWidgetEvent $event): void
     {
-        if (('tl_metamodel_notelist' !== $event->getEnvironment()->getDataDefinition()->getName())
-            || ($event->getProperty()->getName() !== 'name')) {
+        $environment = $event->getEnvironment();
+        if (
+            !(($dataDefinition = $environment->getDataDefinition()) instanceof ContainerInterface)
+            || ('tl_metamodel_notelist' !== $dataDefinition->getName())
+            || ($event->getProperty()->getName() !== 'name')
+        ) {
             return;
         }
 
-        $metaModel = $this->getMetaModelByModelPid($event->getModel());
-        $property  = $event->getProperty();
+        $metaModel     = $this->getMetaModelByModelPid($event->getModel());
+        $property      = $event->getProperty();
+        $languageLabel = $this->translator->trans('name_langcode.label', [], 'tl_metamodel_notelist');
+        $valueLabel    = $this->translator->trans('name_value.label', [], 'tl_metamodel_notelist');
+        $values        =
+            StringUtil::deserialize($event->getModel()->getProperty($event->getProperty()->getName()), true);
 
-        if (!$metaModel->isTranslated()) {
-            $extra = $property->getExtra();
-
-            $extra['tl_class'] .= 'w50';
-
-            $property
-                ->setWidgetType('text')
-                ->setExtra($extra);
-
-            return;
-        }
-
-        $values     = deserialize($event->getModel()->getProperty($event->getProperty()->getName()), true);
-        $fallback   = $metaModel->getFallbackLanguage();
-        $languages  = $this->buildLanguageArray($metaModel, $this->translator);
-        $neededKeys = array_keys($languages);
-
-        // Ensure we have values for all languages present.
-        if (array_diff_key(array_keys($values), $neededKeys)) {
-            foreach ($neededKeys as $langCode) {
-                $values[$langCode] = '';
-            }
-        }
-
-        $rowClasses = [];
-        foreach (array_keys($values) as $langCode) {
-            $rowClasses[] = ($langCode == $fallback) ? 'fallback_language' : 'normal_language';
-        }
-
-        $languageLabel = $this->translator->translate('name_langcode', 'tl_metamodel_notelist');
-        $valueLabel    = $this->translator->translate('name_value', 'tl_metamodel_notelist');
-
-        $extra                   = $property->getExtra();
-        $extra['minCount']       =
-        $extra['maxCount']       = count($languages);
-        $extra['disableSorting'] = true;
-        $extra['tl_class']       = 'clr';
-        $extra['columnFields']   = [
-            'langcode' => [
-                'label'                 => $languageLabel,
-                'exclude'               => true,
-                'inputType'             => 'justtextoption',
-                'options'               => $languages,
-                'eval'                  => [
-                    'rowClasses'        => $rowClasses,
-                    'valign'            => 'center',
-                    'style'             => 'min-width:75px;display:block;'
-                ]
-            ],
-            'value' => [
-                'label'                 => $valueLabel,
-                'exclude'               => true,
-                'inputType'             => 'text',
-                'eval'                  => [
-                    'rowClasses'        => $rowClasses,
-                    'style'             => 'width:400px;',
-                    'rows'              => 3
-                ]
-            ],
-        ];
-
-        $property
-            ->setWidgetType('multiColumnWizard')
-            ->setExtra($extra);
+        Helper::prepareLanguageAwareWidget(
+            $environment,
+            $property,
+            $metaModel,
+            $languageLabel,
+            $valueLabel,
+            false,
+            $values
+        );
     }
 
     /**
@@ -230,7 +153,7 @@ class BuildNoteListNameWidgetListener
      *
      * @throws \InvalidArgumentException When the MetaModel could not be retrieved.
      */
-    private function getMetaModelByModelPid($model)
+    private function getMetaModelByModelPid(ModelInterface $model): IMetaModel
     {
         $metaModel = $this
             ->factory
@@ -241,25 +164,5 @@ class BuildNoteListNameWidgetListener
         }
 
         return $metaModel;
-    }
-
-    /**
-     * Extract all languages from the MetaModel and return them as array.
-     *
-     * @param IMetaModel          $metaModel  The MetaModel to extract the languages from.
-     *
-     * @param TranslatorInterface $translator The translator to use.
-     *
-     * @return \string[]
-     */
-    private function buildLanguageArray(IMetaModel $metaModel, TranslatorInterface $translator)
-    {
-        $languages = array();
-        foreach ((array) $metaModel->getAvailableLanguages() as $langCode) {
-            $languages[$langCode] = $translator->translate('LNG.' . $langCode, 'languages');
-        }
-        asort($languages);
-
-        return $languages;
     }
 }
