@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,12 +12,13 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\EventListener\DcGeneral;
 
@@ -26,11 +27,12 @@ use ContaoCommunityAlliance\Contao\Bindings\Events\Backend\AddToUrlEvent;
 use ContaoCommunityAlliance\Contao\Bindings\Events\Image\GenerateHtmlEvent;
 use ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\Event\ModelToLabelEvent;
 use ContaoCommunityAlliance\DcGeneral\Data\ModelInterface;
-use ContaoCommunityAlliance\Translator\TranslatorInterface;
+use Contao\StringUtil;
 use Doctrine\DBAL\Connection;
 use MetaModels\IFactory;
 use MetaModels\NoteListBundle\NoteListFactory;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * Handles rendering of model from tl_metamodel_filtersetting.
@@ -44,35 +46,35 @@ class FilterSettingTypeRenderer
      *
      * @var TranslatorInterface
      */
-    private $translator;
+    private TranslatorInterface $translator;
 
     /**
      * The event dispatcher.
      *
      * @var EventDispatcherInterface
      */
-    private $dispatcher;
+    private EventDispatcherInterface $dispatcher;
 
     /**
      * The note list factory.
      *
      * @var NoteListFactory
      */
-    private $noteListFactory;
+    private NoteListFactory $noteListFactory;
 
     /**
      * The MetaModels factory.
      *
      * @var IFactory
      */
-    private $factory;
+    private IFactory $factory;
 
     /**
      * The database connection.
      *
      * @var Connection
      */
-    private $connection;
+    private Connection $connection;
 
     /**
      * Create a new instance.
@@ -108,14 +110,15 @@ class FilterSettingTypeRenderer
     {
         $model = $event->getModel();
 
-        if (($model->getProviderName() !== 'tl_metamodel_filtersetting')
+        if (
+            ($model->getProviderName() !== 'tl_metamodel_filtersetting')
             || ('notelist' !== $event->getModel()->getProperty('type'))
         ) {
             return;
         }
 
         $event
-            ->setLabel($this->translator->translate('typedesc.notelist', 'tl_metamodel_filtersetting'))
+            ->setLabel($this->translator->trans('typedesc.notelist', [], 'tl_metamodel_filtersetting'))
             ->setArgs($this->getLabelParameters($model));
     }
 
@@ -126,23 +129,25 @@ class FilterSettingTypeRenderer
      *
      * @return array
      */
-    private function getLabelParameters(ModelInterface $model)
+    private function getLabelParameters(ModelInterface $model): array
     {
         $metaModel = $this->getMetaModel(
-            $model->getProperty('fid'),
+            (string) $model->getProperty('fid'),
             $this->factory,
             $this->connection
         );
+
         if (null === $metaModel) {
             return [];
         }
+
         $lists = $this->noteListFactory->getConfiguredListsFor($metaModel);
 
         return [
             $this->getLabelImage($model),
             $this->getLabelText(),
+            $lists[$model->getProperty('notelist')],
             $this->getLabelComment($model->getProperty('comment')),
-            $lists[$model->getProperty('notelist')]
         ];
     }
 
@@ -153,7 +158,7 @@ class FilterSettingTypeRenderer
      *
      * @return string
      */
-    private function getLabelImage(ModelInterface $model)
+    private function getLabelImage(ModelInterface $model): string
     {
         $image = !$model->getProperty('enabled')
             ? 'bundles/metamodelsnotelist/images/icons/notelist_1.png'
@@ -161,23 +166,23 @@ class FilterSettingTypeRenderer
 
         /** @var AddToUrlEvent $urlEvent */
         $urlEvent = $this->dispatcher->dispatch(
-            ContaoEvents::BACKEND_ADD_TO_URL,
-            new AddToUrlEvent('act=edit&amp;id='.$model->getId())
+            new AddToUrlEvent('act=edit&amp;id=' . $model->getId()),
+            ContaoEvents::BACKEND_ADD_TO_URL
         );
 
         /** @var GenerateHtmlEvent $imageEvent */
         $imageEvent = $this->dispatcher->dispatch(
-            ContaoEvents::IMAGE_GET_HTML,
             new GenerateHtmlEvent(
                 $image,
-                $this->translator->translate('typedesc.notelist', 'tl_metamodel_filtersetting')
-            )
+                $this->translator->trans('typedesc.notelist', [], 'tl_metamodel_filtersetting')
+            ),
+            ContaoEvents::IMAGE_GET_HTML
         );
 
-        return sprintf(
+        return \sprintf(
             '<a href="%s">%s</a>',
             $urlEvent->getUrl(),
-            $imageEvent->getHtml()
+            $imageEvent->getHtml() ?? ''
         );
     }
 
@@ -186,7 +191,7 @@ class FilterSettingTypeRenderer
      *
      * @return string
      */
-    private function getLabelText()
+    private function getLabelText(): string
     {
         return $this->translate('typenames.notelist', 'note list');
     }
@@ -198,14 +203,16 @@ class FilterSettingTypeRenderer
      *
      * @return string
      */
-    private function getLabelComment(string $comment)
+    private function getLabelComment(string $comment): string
     {
         if (!empty($comment)) {
-            return sprintf(
-                $this->translator->translate('typedesc._comment_', 'tl_metamodel_filtersetting'),
-                specialchars($comment)
+            return $this->translator->trans(
+                'typedesc._comment_',
+                ['%comment%' => StringUtil::specialchars($comment)],
+                'tl_metamodel_filtersetting'
             );
         }
+
         return '';
     }
 
@@ -213,17 +220,17 @@ class FilterSettingTypeRenderer
      * Translate a string and return the default if not translated.
      *
      * @param string $string  The language key to translate.
-     *
      * @param string $default The default text to return.
      *
      * @return string
      */
-    private function translate(string $string, string $default = '')
+    private function translate(string $string, string $default = ''): string
     {
-        $label = $this->translator->translate($string, 'tl_metamodel_filtersetting');
-        if ($label == 'typenames.notelist') {
+        $label = $this->translator->trans($string, [], 'tl_metamodel_filtersetting');
+        if ($label === 'typenames.notelist') {
             return $default;
         }
+
         return $label;
     }
 }
