@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 - 2018 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,18 +13,24 @@
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2017 - 2018 The MetaModels team.
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\EventListener;
 
-use Contao\CoreBundle\Exception\RedirectResponseException;
 use ContaoCommunityAlliance\UrlBuilder\UrlBuilder;
 use ContaoCommunityAlliance\UrlBuilder\UrlBuilderFactoryInterface;
+use Contao\ContentModel;
+use Contao\CoreBundle\Exception\RedirectResponseException;
+use Contao\CoreBundle\InsertTag\InsertTagParser;
+use Contao\FormFieldModel;
+use Contao\Model\Collection;
+use Contao\ModuleModel;
+use Contao\System;
 use MetaModels\Events\ParseItemEvent;
 use MetaModels\Events\RenderItemListEvent;
 use MetaModels\FrontendIntegration\HybridList;
@@ -37,57 +43,61 @@ use MetaModels\NoteListBundle\Form\FormBuilder;
 use MetaModels\NoteListBundle\NoteListFactory;
 use MetaModels\NoteListBundle\Storage\NoteListStorage;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * This class adds notelist actions in MetaModel frontend lists.
+ *
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
 class ParseItemListener
 {
     /**
      * Key to use in the render setting.
      */
-    const NOTELIST_LIST = '$note-lists';
+    public const NOTELIST_LIST = '$note-lists';
 
     /**
      * Key to use for flag in the render setting for disabling form rendering.
      */
-    const NOTELIST_LIST_DISABLE_FORM = '$note-lists-no-form';
+    public const NOTELIST_LIST_DISABLE_FORM = '$note-lists-no-form';
 
     /**
      * The note list factory.
      *
      * @var NoteListFactory
      */
-    private $factory;
+    private NoteListFactory $factory;
 
     /**
      * The event dispatcher.
      *
      * @var EventDispatcherInterface
      */
-    private $dispatcher;
+    private EventDispatcherInterface $dispatcher;
 
     /**
      * The form builder.
      *
      * @var FormBuilder
      */
-    private $formBuilder;
+    private FormBuilder $formBuilder;
 
     /**
      * The url builder factory.
      *
      * @var UrlBuilderFactoryInterface
      */
-    private $urlBuilderFactory;
+    private UrlBuilderFactoryInterface $urlBuilderFactory;
 
     /**
      * The request stack.
      *
      * @var RequestStack
      */
-    private $requestStack;
+    private RequestStack $requestStack;
 
     /**
      * Create a new instance.
@@ -121,11 +131,17 @@ class ParseItemListener
      *
      * @SuppressWarnings(PHPMD.Superglobals)
      * @SuppressWarnings(PHPMD.CamelCaseVariableName)
+     *
+     * @psalm-suppress UndefinedMagicPropertyFetch
      */
-    public function handleListRendering(RenderItemListEvent $event)
+    public function handleListRendering(RenderItemListEvent $event): void
     {
         $caller = $event->getCaller();
-        if (!($caller instanceof HybridList)) {
+        if (
+            !($caller instanceof HybridList)
+            && !($caller instanceof ContentModel)
+            && !($caller instanceof ModuleModel)
+        ) {
             return;
         }
 
@@ -133,7 +149,7 @@ class ParseItemListener
             return;
         }
 
-        $lists = !empty($tmp = $caller->metamodel_notelist) ? unserialize($tmp) : [];
+        $lists = null !== ($tmp = $caller->metamodel_notelist) ? \unserialize($tmp, ['allowed_classes' => false]) : [];
 
         if (!$this->processActions($event->getList()->getMetaModel(), $lists)) {
             $event->getList()->getView()->set(self::NOTELIST_LIST, $lists);
@@ -147,7 +163,7 @@ class ParseItemListener
      *
      * @return void
      */
-    public function handleFormRendering(ParseNoteListFormEvent $event)
+    public function handleFormRendering(ParseNoteListFormEvent $event): void
     {
         $renderSetting = $event->getRenderSetting();
 
@@ -166,10 +182,10 @@ class ParseItemListener
      *
      * @return void
      */
-    public function addNoteListActions(ParseItemEvent $event)
+    public function addNoteListActions(ParseItemEvent $event): void
     {
         $settings = $event->getRenderSettings();
-        if (!($lists = $settings->get(self::NOTELIST_LIST))) {
+        if (null === ($lists = $settings->get(self::NOTELIST_LIST))) {
             return;
         }
 
@@ -183,28 +199,28 @@ class ParseItemListener
                 continue;
             }
 
-            // Add notelist name.
+            // Add note list name.
             $parsed['notelists_names']['notelist_' . $list] = $storage->getName();
 
             if ($formId = $storage->getMeta()->get('form')) {
                 // Add payload values.
-                if (count($storage->getMetaDataFor($item))) {
-                    $parsed['notelists_payload_values']['notelist_' . $list] = $storage->getMetaDataFor($item);
+                if (\count($storageData = $storage->getMetaDataFor($item))) {
+                    $parsed['notelists_payload_values']['notelist_' . $list] = $storageData;
                 }
 
                 // Add payload labels.
-                $parsed['notelists_payload_labels']['notelist_' . $list] = $this->getFormFieldLabels(intval($formId));
+                $parsed['notelists_payload_labels']['notelist_' . $list] = $this->getFormFieldLabels((int) $formId);
             }
 
             // Hide all other actions input if form disabled.
-            if ($settings->get(self::NOTELIST_LIST_DISABLE_FORM)) {
+            if (null !== $settings->get(self::NOTELIST_LIST_DISABLE_FORM)) {
                 continue;
             }
 
             if ($formId = $storage->getMeta()->get('form')) {
                 // Need to render the form here.
                 $parsed['actions']['notelist_' . $list . '_form'] =
-                    $this->generateForm($item, $storage, intval($formId));
+                    $this->generateForm($item, $storage, (int) $formId);
                 if (!$storage->has($item)) {
                     continue;
                 }
@@ -225,11 +241,11 @@ class ParseItemListener
      *
      * @throws \InvalidArgumentException When the item could not be found or the action is unknown.
      */
-    private function processActions(IMetaModel $metaModel, array $lists)
+    private function processActions(IMetaModel $metaModel, array $lists): bool
     {
         foreach ($lists as $list) {
             if ($event = $this->buildActionEvent($metaModel, $list)) {
-                $this->dispatcher->dispatch(NoteListEvents::PROCESS_NOTE_LIST_ACTION, $event);
+                $this->dispatcher->dispatch($event, NoteListEvents::PROCESS_NOTE_LIST_ACTION);
                 if ($event->isSuccess()) {
                     $this->redirect($list);
                     return true;
@@ -252,25 +268,25 @@ class ParseItemListener
      *
      * @return ProcessActionEvent|null
      */
-    private function buildActionEvent(IMetaModel $metaModel, string $list)
+    private function buildActionEvent(IMetaModel $metaModel, string $list): ?ProcessActionEvent
     {
         $url = $this->getCurrentUrl();
 
         if ($url->hasQueryParameter('notelist_' . $list . '_action')) {
             return new ProcessActionEvent(
-                $url->getQueryParameter('notelist_' . $list . '_action'),
-                ['item' => $url->getQueryParameter('notelist_' . $list . '_item')],
+                $url->getQueryParameter('notelist_' . $list . '_action') ?? '',
+                ['item' => $url->getQueryParameter('notelist_' . $list . '_item') ?? ''],
                 $this->factory->getList($metaModel, $list),
                 $metaModel
             );
         }
-        if (!($noteList = $this->factory->getList($metaModel, $list))) {
-            return null;
-        }
+
+        $noteList = $this->factory->getList($metaModel, $list);
+
         $valueBag = $noteList->getMeta();
         if ($valueBag->has('form') && ($formId = $valueBag->get('form'))) {
-            $form = $this->formBuilder->getForm(intval($formId), $noteList, $this->getCurrentUrl()->getUrl());
-            if ($data = $form->getSubmittedData()) {
+            $form = $this->formBuilder->getForm((int) $formId, $noteList, $this->getCurrentUrl()->getUrl());
+            if (null !== ($data = $form->getSubmittedData())) {
                 return new ProcessActionEvent(
                     'add',
                     $data,
@@ -279,6 +295,7 @@ class ParseItemListener
                 );
             }
         }
+
         return null;
     }
 
@@ -293,7 +310,7 @@ class ParseItemListener
      * @SuppressWarnings(PHPMD.Superglobals)
      * @SuppressWarnings(PHPMD.CamelCaseVariableName)
      */
-    private function generateButton(IItem $item, NoteListStorage $storage)
+    private function generateButton(IItem $item, NoteListStorage $storage): array
     {
         $action = !$storage->has($item) ? 'add' : 'remove';
         $url    = $this
@@ -302,9 +319,16 @@ class ParseItemListener
             ->setQueryParameter('notelist_' . $storage->getStorageKey() . '_item', $item->get('id'));
 
         // Obtain list and generate button for it.
+        $translator = System::getContainer()->get('translator');
+        assert($translator instanceof TranslatorInterface);
+
         return [
             'name'  => $storage->getName(),
-            'label' => sprintf($GLOBALS['TL_LANG']['MSC']['metamodel_notelist_' . $action], $storage->getName()),
+            'label' => $translator->trans(
+                'metamodel_notelist.' . $action,
+                ['%list_name%' => $storage->getName()],
+                'notelist_default'
+            ),
             'href'  => $url->getUrl(),
             'class' => $action,
             'meta'  => $storage->getMetaDataFor($item),
@@ -320,7 +344,7 @@ class ParseItemListener
      *
      * @return array
      */
-    private function generateForm(IItem $item, NoteListStorage $storage, int $formId)
+    private function generateForm(IItem $item, NoteListStorage $storage, int $formId): array
     {
         $form = $this->formBuilder->getForm($formId, $storage, $this->getCurrentUrl()->getUrl());
 
@@ -334,15 +358,18 @@ class ParseItemListener
      *
      * @return array
      */
-    private function getFormFieldLabels(int $formId)
+    private function getFormFieldLabels(int $formId): array
     {
         $formLabels = [];
 
         if ($formId) {
-            $objFields = \FormFieldModel::findPublishedByPid($formId);
+            $objFields = FormFieldModel::findPublishedByPid($formId);
+            assert($objFields instanceof Collection);
+            $parser    = System::getContainer()->get('contao.insert_tag.parser');
+            assert($parser instanceof InsertTagParser);
 
             foreach ($objFields as $objField) {
-                $formLabels[$objField->name] = \Controller::replaceInsertTags($objField->label, false);
+                $formLabels[$objField->name] = $parser->replace($objField->label);
             }
         }
 
@@ -358,7 +385,7 @@ class ParseItemListener
      *
      * @throws RedirectResponseException In order to redirect.
      */
-    private function redirect(string $identifier)
+    private function redirect(string $identifier): void
     {
         $url = $this
             ->getCurrentUrl()
@@ -370,15 +397,16 @@ class ParseItemListener
     }
 
     /**
-     * Retrieve an URL builder containing the current URL.
+     * Retrieve a URL builder containing the current URL.
      *
      * @return UrlBuilder
      *
      * @internal
      */
-    protected function getCurrentUrl()
+    protected function getCurrentUrl(): UrlBuilder
     {
         $request = $this->requestStack->getCurrentRequest();
+        assert($request instanceof Request);
 
         return $this->urlBuilderFactory->create($request->getRequestUri());
     }
