@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,12 +13,12 @@
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2017 The MetaModels team.
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle;
 
@@ -34,14 +34,14 @@ class InsertTags
      *
      * @var NoteListFactory
      */
-    private $factory;
+    private NoteListFactory $factory;
 
     /**
      * The MetaModel factory.
      *
      * @var IFactory
      */
-    private $metaModelFactory;
+    private IFactory $metaModelFactory;
 
     /**
      * Create a new instance.
@@ -70,9 +70,9 @@ class InsertTags
         }
 
         if (!empty($arguments[1])) {
-            $lists = explode(',', $arguments[1]);
+            $lists = \explode(',', $arguments[1]);
         } else {
-            $lists = array_keys($this->factory->getConfiguredListsFor($metaModel));
+            $lists = \array_keys($this->factory->getConfiguredListsFor($metaModel));
         }
 
         $sum = 0;
