@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,12 +13,12 @@
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2017 The MetaModels team.
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\Storage;
 
@@ -35,7 +35,7 @@ class StorageAdapterFactory
      *
      * @var AdapterInterface[]
      */
-    private $instances = [];
+    private array $instances = [];
 
     /**
      * Retrieve an adapter by name.
@@ -52,6 +52,7 @@ class StorageAdapterFactory
             return $this->instances[$identifier];
         }
 
+        /** @psalm-suppress DeprecatedClass */
         switch ($identifier) {
             case 'php-session':
                 return $this->instances[$identifier] = new PhpSessionVariableAdapter();
