@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,12 +12,13 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\Event;
 
@@ -29,12 +30,12 @@ class NoteListEvents
     /**
      * This event gets triggered when a note list gets rendered in a frontend form.
      */
-    const PARSE_NOTE_LIST_FORM = 'metamodels.note-list.parse-note-list-form';
+    public const PARSE_NOTE_LIST_FORM = 'metamodels.note-list.parse-note-list-form';
 
     /**
      * This event gets triggered when a note list gets manipulated.
      */
-    const MANIPULATE_NOTE_LIST = 'metamodels.note-list.manipulate';
+    public const MANIPULATE_NOTE_LIST = 'metamodels.note-list.manipulate';
 
-    const PROCESS_NOTE_LIST_ACTION = 'metamodels.note-list.process-action';
+    public const PROCESS_NOTE_LIST_ACTION = 'metamodels.note-list.process-action';
 }
