@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,17 +12,20 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\EventListener\DcGeneral;
 
+use Contao\StringUtil;
 use ContaoCommunityAlliance\DcGeneral\Contao\View\Contao2BackendView\Event\GetBreadcrumbEvent;
 use ContaoCommunityAlliance\DcGeneral\Data\ModelId;
+use ContaoCommunityAlliance\DcGeneral\DataDefinition\ContainerInterface;
 use ContaoCommunityAlliance\DcGeneral\EnvironmentInterface;
 use ContaoCommunityAlliance\UrlBuilder\UrlBuilder;
 use MetaModels\CoreBundle\EventListener\DcGeneral\Breadcrumb\AbstractBreadcrumbListener;
@@ -39,9 +42,12 @@ class BreadCrumbNoteList extends AbstractBreadcrumbListener
     /**
      * {@inheritDoc}
      */
-    protected function wantToHandle(GetBreadcrumbEvent $event)
+    protected function wantToHandle(GetBreadcrumbEvent $event): bool
     {
-        return 'tl_metamodel_notelist' === $event->getEnvironment()->getDataDefinition()->getName();
+        $container = $event->getEnvironment()->getDataDefinition();
+        assert($container instanceof ContainerInterface);
+
+        return 'tl_metamodel_notelist' === $container->getName();
     }
 
     /**
@@ -69,10 +75,10 @@ class BreadCrumbNoteList extends AbstractBreadcrumbListener
             ->unsetQueryParameter('id');
 
         $elements->push(
-            ampersand($builder->getUrl()),
-            sprintf(
+            StringUtil::ampersand($builder->getUrl()),
+            \sprintf(
                 $elements->getLabel('tl_metamodel_notelist'),
-                $this->getMetaModel($elements->getId('tl_metamodel'))->getName()
+                $this->getMetaModel($elements->getId('tl_metamodel') ?? '')->getName()
             ),
             'bundles/metamodelsnotelist/images/icons/notelist.png'
         );
