@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,17 +12,19 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\EventListener\DcGeneral;
 
 use Doctrine\DBAL\Connection;
 use MetaModels\IFactory;
+use MetaModels\IMetaModel;
 
 /**
  * This trait converts a 'fid' value from tl_metamodel_filtersetting to a MetaModel instance.
@@ -38,9 +40,9 @@ trait FilterIdToMetaModelTrait
      * @param IFactory   $factory    The MetaModels factory.
      * @param Connection $connection The database connection.
      *
-     * @return \MetaModels\IMetaModel|null
+     * @return IMetaModel|null
      */
-    private function getMetaModel($fid, IFactory $factory, Connection $connection)
+    private function getMetaModel(string $fid, IFactory $factory, Connection $connection): ?IMetaModel
     {
         // This is pretty lame and hardcoded - we need to adjust this when we have non DB based definitions.
         $filter = $connection
@@ -49,18 +51,20 @@ trait FilterIdToMetaModelTrait
             ->from('tl_metamodel_filter')
             ->where('id=:id')
             ->setParameter('id', $fid)
-            ->execute()
-            ->fetch(\PDO::FETCH_ASSOC);
+            ->executeQuery()
+            ->fetchAssociative();
+
         if (false === $filter) {
             return null;
         }
 
-        if (null === ($metaModelName = $factory->translateIdToMetaModelName($filter['pid']))) {
+        if ('' === ($metaModelName = $factory->translateIdToMetaModelName($filter['pid']))) {
             return null;
         }
         if (null === ($metaModel = $factory->getMetaModel($metaModelName))) {
             return null;
         }
+
         return $metaModel;
     }
 }
