@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,12 +12,13 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\EventListener\DcGeneral;
 
@@ -25,6 +26,8 @@ use ContaoCommunityAlliance\DcGeneral\View\Event\RenderReadablePropertyValueEven
 
 /**
  * This class renders the note list name as readable property value for dc-general.
+ *
+ * @SuppressWarnings(PHPMD.LongClassName)
  */
 class RenderNoteListNameAsReadablePropertyValueListener
 {
@@ -40,9 +43,11 @@ class RenderNoteListNameAsReadablePropertyValueListener
      */
     public function render(RenderReadablePropertyValueEvent $event)
     {
-        if (null !== $event->getRendered()
+        if (
+            null !== $event->getRendered()
             || 'tl_metamodel_notelist' !== $event->getModel()->getProviderName()
-            || $event->getProperty()->getName() !== 'name') {
+            || $event->getProperty()->getName() !== 'name'
+        ) {
             return;
         }
 
