@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,12 +13,12 @@
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2017 The MetaModels team.
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\EventListener;
 
@@ -69,7 +69,7 @@ class ProcessActionListener
      */
     private function getItemFromMetaModel(ProcessActionEvent $event)
     {
-        $item = $event->getMetaModel()->findById($event->getPayloadValue('item'));
+        $item = $event->getMetaModel()->findById($event->getPayloadValue('item') ?? '');
         if (null === $item) {
             throw new \InvalidArgumentException('Item ' . $event->getPayloadValue('item') . ' could not be found.');
         }
