@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,18 +12,22 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace MetaModels\NoteListBundle\Form;
 
 use Contao\Controller;
 use Contao\FormFieldModel;
+use Contao\FormHidden;
 use Contao\FormModel;
+use Contao\Model\Collection;
+use Contao\StringUtil;
 use Contao\Widget;
 use MetaModels\NoteListBundle\Storage\NoteListStorage;
 
@@ -41,7 +45,7 @@ class FormBuilder
      *
      * @return Form
      */
-    public function getForm(int $formId, NoteListStorage $noteList, string $action)
+    public function getForm(int $formId, NoteListStorage $noteList, string $action): Form
     {
         return new Form(
             $this->getFormConfig($formId),
@@ -56,11 +60,14 @@ class FormBuilder
      *
      * @param int $formId The form ID.
      *
-     * @return Form[]
+     * @return FormModel
      */
-    private function getFormConfig($formId)
+    private function getFormConfig(int $formId): FormModel
     {
-        return FormModel::findById($formId);
+        $form = FormModel::findById($formId);
+        assert($form instanceof FormModel);
+
+        return $form;
     }
 
     /**
@@ -70,22 +77,26 @@ class FormBuilder
      *
      * @return Widget[]
      */
-    private function getFormWidgets($formId)
+    private function getFormWidgets(int $formId): array
     {
         Controller::loadDataContainer('tl_form_field');
         // Get all form fields
-        if (null === ($objFields = FormFieldModel::findPublishedByPid($formId)) || 0 === $objFields->count()) {
+        $objFields = FormFieldModel::findPublishedByPid($formId);
+        if (!($objFields instanceof Collection) || 0 === $objFields->count()) {
             return [];
         }
         $fields = [];
         $hidden = [];
 
-        // Process the fields
+        // Process the fields.
         $row    = 0;
-        $maxRow = count($fields);
+        $maxRow = \count($fields);
         foreach ($objFields as $objField) {
             $widget = $this->buildWidget($objField, $row, $maxRow);
-            if ($widget instanceof \FormHidden) {
+            if (null === $widget) {
+                continue;
+            }
+            if ($widget instanceof FormHidden) {
                 $hidden[] = $widget;
                 --$maxRow;
                 continue;
@@ -95,7 +106,7 @@ class FormBuilder
             ++$row;
         }
 
-        return array_merge($hidden, $fields);
+        return \array_merge($hidden, $fields);
     }
 
     /**
@@ -110,10 +121,10 @@ class FormBuilder
      * @SuppressWarnings(PHPMD.Superglobals)
      * @SuppressWarnings(PHPMD.CamelCaseVariableName)
      */
-    private function buildWidget(FormFieldModel $field, int &$row, int &$maxRow)
+    private function buildWidget(FormFieldModel $field, int &$row, int &$maxRow): Widget|\Widget|null
     {
         // Continue if the class is not defined
-        if (!class_exists($strClass = $GLOBALS['TL_FFL'][$field->type])) {
+        if (!\class_exists($strClass = ($GLOBALS['TL_FFL'][$field->type] ?? ''))) {
             return null;
         }
 
@@ -122,21 +133,27 @@ class FormBuilder
         $arrData['allowHtml']      = true;
         $arrData['rowClass']       = $this->getWidgetClass($row, $maxRow);
         $arrData['tableless']      = true;
-        // Increase the row count if its a password field
-        if ($field->type == 'password') {
+
+        // Increase the row count if it's a password field
+        if ($field->type === 'password') {
             ++$row;
             ++$maxRow;
             $arrData['rowClassConfirm'] = $this->getWidgetClass($row, $maxRow);
         }
 
         // Submit buttons do not use the name attribute
-        if ($field->type == 'submit') {
+        if ($field->type === 'submit') {
             $arrData['name'] = '';
         }
 
         // Unset the default value depending on the field type (see #4722)
         if (!empty($arrData['value'])) {
-            if (!in_array('value', trimsplit('[,;]', $GLOBALS['TL_DCA']['tl_form_field']['palettes'][$field->type]))) {
+            if (
+                !\in_array(
+                    'value',
+                    StringUtil::trimsplit('[,;]', $GLOBALS['TL_DCA']['tl_form_field']['palettes'][$field->type])
+                )
+            ) {
                 $arrData['value'] = '';
             }
         }
@@ -156,9 +173,9 @@ class FormBuilder
      *
      * @return string
      */
-    private function getWidgetClass(int $row, int $last)
+    private function getWidgetClass(int $row, int $last): string
     {
-        $class = 'row_'.$row;
+        $class = 'row_' . $row;
         if (0 === $row) {
             $class .= ' row_first';
         }
