@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,14 +12,16 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
 
 $GLOBALS['TL_DCA']['tl_metamodel']['list']['operations']['notelist'] = [
-    'label'   => &$GLOBALS['TL_LANG']['tl_metamodel']['notelist'],
-    'href'    => 'table=tl_metamodel_notelist',
-    'icon'    => 'bundles/metamodelsnotelist/images/icons/notelist.png',
-    'idparam' => 'pid'
+    'label'       => 'notelist.label',
+    'description' => 'notelist.description',
+    'href'        => 'table=tl_metamodel_notelist',
+    'icon'        => 'bundles/metamodelsnotelist/images/icons/notelist.png',
+    'idparam'     => 'pid'
 ];
