@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/notelist.
  *
- * (c) 2017 The MetaModels team.
+ * (c) 2017-2025 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,7 +12,8 @@
  *
  * @package    MetaModels
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2017 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2017-2025 The MetaModels team.
  * @license    https://github.com/MetaModels/notelist/blob/master/LICENSE LGPL-3.0
  * @filesource
  */
@@ -24,13 +25,14 @@ $GLOBALS['TL_DCA']['tl_metamodel_filtersetting']['metapalettes']['notelist exten
 ];
 
 $GLOBALS['TL_DCA']['tl_metamodel_filtersetting']['fields']['notelist'] = [
-    'label'                   => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['notelist'],
-    'exclude'                 => true,
-    'default'                 => true,
-    'inputType'               => 'select',
-    'sql'                     => 'int(10) unsigned NOT NULL default \'0\'',
-    'eval'                    => [
-        'tl_class'            => 'w50',
+    'label'       => 'notelist.label',
+    'description' => 'notelist.description',
+    'exclude'     => true,
+    'default'     => true,
+    'inputType'   => 'select',
+    'sql'         => 'int(10) unsigned NOT NULL default \'0\'',
+    'eval'        => [
+        'tl_class'           => 'w50',
         'includeBlankOption' => true,
         'mandatory'          => true,
     ],
